services/scenario-management/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "strawberry-graphql[fastapi]>=0.278.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
